+++ detectors/heroku/example_test.go
@@ -1,0 +1,28 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package heroku_test
+
+import (
+	"context"
+	"log"
+
+	"go.opentelemetry.io/otel/sdk/resource"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+
+	"go.opentelemetry.io/contrib/detectors/heroku"
+)
+
+func ExampleNewResourceDetector() {
+	res, err := resource.New(
+		context.Background(),
+		resource.WithDetectors(heroku.NewResourceDetector()),
+	)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	tp := sdktrace.NewTracerProvider(sdktrace.WithResource(res))
+	_ = tp.Shutdown(context.Background())
+	// Use tp to create tracers ...
+}